HomeView分割でtype_body_length解消
